requirements.txt
the-new-hotness
fastapi
uvicorn
streamlit
openai
python-dotenv
sentence-transformers
faiss-cpu
pydantic
requests
